backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
python-dotenv
aiofiles
pydantic
requests
whisper
